build-gcc.sh: fixes path
[BETA]: Since we have this script :https://github.com/Pablito2020/pablito-scripts/blob/master/daredevil.sh in my pablito-scripts project we have to think that the people who uses it, have already renamed the kernel directory so... I thought this:
if the daredevil.sh script that is in the Daredevil-Kernel folder doesn't exist,
then rename the android_kernel_bq_krillin folder to Daredevil-Kernel and enter to it.

And that's it! Is in beta version because I haven't tried it yet so... Hope it works!

PS: This commit will be applied to the build-linaro.sh too

+++ Daredevil-Scripts/build-gcc.sh
@@ -53,9 +53,13 @@ then
     echo -e "####################################"
 cd ..
 git clone https://android.googlesource.com/platform/prebuilts/gcc/linux-x86/arm/arm-eabi-4.8
-# Rename the folder android_kernel_bq_krillin (git project) to Daredevil-Kernel and enter to it
-mv android_kernel_bq_krillin Daredevil-Kernel
-cd Daredevil-Kernel
+if [ ! -f Daredevil-Kernel/daredevil.sh ]
+then
+    mv android_kernel_bq_krillin Daredevil-Kernel
+    cd Daredevil-Kernel
+else
+    cd Daredevil-Kernel
+fi
 export ARCH=arm CROSS_COMPILE=../arm-eabi-4.8/bin/arm-eabi-
 else
 export ARCH=arm CROSS_COMPILE=../arm-eabi-4.8/bin/arm-eabi-
